+++ angular-architecture/src/app/core/core-routing.module.ts
@@ -21,13 +21,6 @@ const routes: Routes = [{
   ]
 }];
 
-// { path: 'dashboard', loadChildren: './dashboard/dashboard.module#DashboardModule' },
-// { path: 'page-not-found', loadChildren: './not-found/not-found.module#NotFoundModule' },
-// { path: 'internal-error', loadChildren: './internal-error/internal-error.module#InternalErrorModule' },
-
-// { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
-// { path: '**', redirectTo: 'page-not-found', pathMatch: 'full' }
-
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]